WebUI/requirements.txt
Flask==3.1.0
protobuf==6.30.0
python-dotenv==1.0.1
